% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dietcheck.R
\docType{data}
\name{dietcheck}
\alias{dietcheck}
\title{Predation on salmon individual scenarios}
\format{
A data frame with 2071296 rows and 6 variables:
\describe{
\item{\code{Time}}{double days}
\item{\code{Predator}}{character predator code}
\item{\code{model_ver}}{integer ensemble model version}
\item{\code{scenario_name}}{character scenario name}
\item{\code{scenario_var}}{double scenario variation}
\item{\code{tot_mort}}{double total salmon mortality}
}
}
\usage{
dietcheck
}
\description{
Predation based on dietcheck, aggregated by salmon predator
}
\details{
DETAILS
}
\keyword{datasets}
