% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dietcheckcum.R
\docType{data}
\name{dietcheckcum}
\alias{dietcheckcum}
\title{DATASET_TITLE}
\format{
A data frame with 691200 rows and 6 variables:
\describe{
\item{\code{Time}}{double COLUMN_DESCRIPTION}
\item{\code{Predator}}{character COLUMN_DESCRIPTION}
\item{\code{model_ver}}{integer COLUMN_DESCRIPTION}
\item{\code{scenario_name}}{character COLUMN_DESCRIPTION}
\item{\code{scenario_var}}{character COLUMN_DESCRIPTION}
\item{\code{tot_mort}}{double COLUMN_DESCRIPTION}
}
}
\usage{
dietcheckcum
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
