% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ensemblebiomass.R
\docType{data}
\name{ensemblebiomass}
\alias{ensemblebiomass}
\title{A dataset of biomass of 30 yr ensemble simulations}
\format{
A data.frame with . rows and . variables:
\describe{
\item{\code{Code}}{Functional group code; character.}
\item{\code{biomass}}{Biomass in metric tons; number.}
\item{\code{name}}{Functional group name; character.}
\item{\code{longname}}{Functional group long name; character.}
\item{\code{Year}}{Simulation year; number.}
\item{\code{model_ver}}{Model variant; integer.}
}

A data frame.
}
\usage{
data(ensemblebiomass)
}
\description{
Biomass of 8 model variants, under mfc fishing
}
\details{
This dataset in the output of simulations of each model variant.
}
\references{
Simulation data.
}
\keyword{datasets}
