% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ensemblebiomasscum.R
\docType{data}
\name{ensemblebiomasscum}
\alias{ensemblebiomasscum}
\title{DATASET_TITLE}
\format{
A data frame with 536352 rows and 8 variables:
\describe{
\item{\code{Code}}{character COLUMN_DESCRIPTION}
\item{\code{biomass}}{double COLUMN_DESCRIPTION}
\item{\code{name}}{character COLUMN_DESCRIPTION}
\item{\code{longname}}{character COLUMN_DESCRIPTION}
\item{\code{Year}}{double COLUMN_DESCRIPTION}
\item{\code{model_ver}}{integer COLUMN_DESCRIPTION}
\item{\code{scenario_name}}{character COLUMN_DESCRIPTION}
\item{\code{scenario_var}}{character COLUMN_DESCRIPTION}
}
}
\usage{
ensemblebiomasscum
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
