% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ensemblenumbersage.R
\docType{data}
\name{ensemblenumbersage}
\alias{ensemblenumbersage}
\title{A data set of salmon group numbers at age of 30 yr ensemble simulations}
\format{
A data frame with 117600 rows and 12 variables:
\describe{
\item{\code{nums}}{double; numbers at age}
\item{\code{agename}}{character; functional group age class name}
\item{\code{group}}{character; functional group name}
\item{\code{age}}{integer; functional group age}
\item{\code{Code}}{character; functional group code}
\item{\code{years_away}}{integer; years salmon spend at sea}
\item{\code{Long.Name}}{character; long functional group name}
\item{\code{NumCohorts}}{integer; functional group number of cohorts}
\item{\code{timestep}}{integer; simulation timestep}
\item{\code{year}}{double; simulation year number}
\item{\code{year_no}}{integer; simulation year}
\item{\code{model_ver}}{integer; model number}
}
}
\usage{
ensemblenumbersage
}
\description{
Salmon numbers at age for 8 model variants to plot survival
}
\details{
DETAILS
}
\keyword{datasets}
