% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ensemblenumbersagescenarios.R
\docType{data}
\name{ensemblenumbersagescenarios}
\alias{ensemblenumbersagescenarios}
\title{DATASET_TITLE}
\format{
A data frame with 2102100 rows and 14 variables:
\describe{
\item{\code{nums}}{double; numbers at age}
\item{\code{agename}}{character; functional group age class name}
\item{\code{group}}{character; functional group name}
\item{\code{age}}{integer; functional group age}
\item{\code{Code}}{character; functional group code}
\item{\code{years_away}}{integer; years salmon spend at sea}
\item{\code{longname}}{character; long functional group name}
\item{\code{NumCohorts}}{integer; functional group number of cohorts}
\item{\code{timestep}}{integer; simulation timestep}
\item{\code{year}}{double; simulation year number}
\item{\code{year_no}}{integer; simulation year}
\item{\code{model_ver}}{integer; model number}
\item{\code{scenario_name}}{character salmon survival scenario}
\item{\code{scenario_var}}{character model variant}
}
}
\usage{
ensemblenumbersagescenarios
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
