% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ensemblenumberscum.R
\docType{data}
\name{ensemblenumberscum}
\alias{ensemblenumberscum}
\title{DATASET_TITLE}
\format{
A data frame with 705600 rows and 14 variables:
\describe{
\item{\code{nums}}{double COLUMN_DESCRIPTION}
\item{\code{agename}}{character COLUMN_DESCRIPTION}
\item{\code{group}}{character COLUMN_DESCRIPTION}
\item{\code{age}}{integer COLUMN_DESCRIPTION}
\item{\code{Code}}{character COLUMN_DESCRIPTION}
\item{\code{years_away}}{integer COLUMN_DESCRIPTION}
\item{\code{longname}}{character COLUMN_DESCRIPTION}
\item{\code{NumCohorts}}{integer COLUMN_DESCRIPTION}
\item{\code{timestep}}{integer COLUMN_DESCRIPTION}
\item{\code{year}}{double COLUMN_DESCRIPTION}
\item{\code{year_no}}{integer COLUMN_DESCRIPTION}
\item{\code{model_ver}}{integer COLUMN_DESCRIPTION}
\item{\code{scenario_name}}{character COLUMN_DESCRIPTION}
\item{\code{scenario_var}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
ensemblenumberscum
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
