% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ensemblewagecum.R
\docType{data}
\name{ensemblewagecum}
\alias{ensemblewagecum}
\title{Weight-at-age in cumulative scenarios}
\format{
A data frame with 183456 rows and 11 variables:
\describe{
\item{\code{variable}}{double weight-at-age}
\item{\code{time}}{integer timestep}
\item{\code{age}}{integer age}
\item{\code{group}}{character group}
\item{\code{variable_type}}{character COLUMN_DESCRIPTION}
\item{\code{code}}{character functional group code}
\item{\code{name}}{character functional group name}
\item{\code{model_ver}}{integer ensemble member}
\item{\code{scenario_name}}{character scenario name}
\item{\code{scenario_var}}{character salmon effect}
\item{\code{year}}{double year}
}
}
\usage{
ensemblewagecum
}
\description{
Weight-at-age for cumulative scenarios
}
\details{
DETAILS
}
\keyword{datasets}
