% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ensemblewageind.R
\docType{data}
\name{ensemblewageind}
\alias{ensemblewageind}
\title{Weight-at-age in individual scenarios}
\format{
An object of class \code{data.frame} with 547820 rows and 11 columns.
}
\usage{
ensemblewageind
}
\description{
Weight-at-age for individual scenarios
#' @format A data frame with 183456 rows and 11 variables:
\describe{
\item{\code{variable}}{double weight-at-age}
\item{\code{time}}{integer timestep}
\item{\code{age}}{integer age}
\item{\code{group}}{character group}
\item{\code{variable_type}}{character COLUMN_DESCRIPTION}
\item{\code{code}}{character functional group code}
\item{\code{name}}{character functional group name}
\item{\code{model_ver}}{integer ensemble member}
\item{\code{scenario_name}}{character scenario name}
\item{\code{scenario_var}}{character salmon effect}
\item{\code{year}}{double year}
}
}
\details{
DETAILS
}
\keyword{datasets}
