% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_base_cum_survival.R
\name{extract_base_cum_survival}
\alias{extract_base_cum_survival}
\title{plot salmon survival of cumulative scenarios of ensemble models}
\usage{
extract_base_cum_survival(ensemblenumberscum, salmongroups)
}
\arguments{
\item{ensemblenumbersage, }{a data frame}

\item{func.groups, }{a data frame of salmon groups in the model}
}
\value{
salmon.return.nums, survival over time
}
\description{
Code to plot survival in cumulative scenarios of multiple AMPS versions
}
\author{
Hem Nalini Morzaria-Luna, hmorzarialuna_gmail.com October 2022
}
