% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-functionalgroups.R
\docType{data}
\name{functionalgroups}
\alias{functionalgroups}
\title{DATASET_TITLE}
\format{
A data frame with 74 rows and 30 variables:
\describe{
\item{\code{Code}}{character COLUMN_DESCRIPTION}
\item{\code{Index}}{integer COLUMN_DESCRIPTION}
\item{\code{IsTurnedOn}}{integer COLUMN_DESCRIPTION}
\item{\code{name}}{character COLUMN_DESCRIPTION}
\item{\code{longname}}{character COLUMN_DESCRIPTION}
\item{\code{NumCohorts}}{integer COLUMN_DESCRIPTION}
\item{\code{NumGeneTypes}}{integer COLUMN_DESCRIPTION}
\item{\code{NumStages}}{integer COLUMN_DESCRIPTION}
\item{\code{NumSpawns}}{integer COLUMN_DESCRIPTION}
\item{\code{NumAgeClassSize}}{double COLUMN_DESCRIPTION}
\item{\code{NumStocks}}{integer COLUMN_DESCRIPTION}
\item{\code{VerticallyMigrates}}{integer COLUMN_DESCRIPTION}
\item{\code{HorizontallyMigrates}}{integer COLUMN_DESCRIPTION}
\item{\code{IsFished}}{integer COLUMN_DESCRIPTION}
\item{\code{IsImpacted}}{integer COLUMN_DESCRIPTION}
\item{\code{isTAC}}{integer COLUMN_DESCRIPTION}
\item{\code{GroupType}}{character COLUMN_DESCRIPTION}
\item{\code{IsPredator}}{integer COLUMN_DESCRIPTION}
\item{\code{IsCover}}{integer COLUMN_DESCRIPTION}
\item{\code{IsSiliconDep}}{integer COLUMN_DESCRIPTION}
\item{\code{IsAssessed}}{integer COLUMN_DESCRIPTION}
\item{\code{IsCatchGrazer}}{integer COLUMN_DESCRIPTION}
\item{\code{OverWinters}}{integer COLUMN_DESCRIPTION}
\item{\code{isCultured}}{integer COLUMN_DESCRIPTION}
\item{\code{isHabDepend}}{integer COLUMN_DESCRIPTION}
\item{\code{numMoveEntries}}{integer COLUMN_DESCRIPTION}
\item{\code{isSupplemented}}{integer COLUMN_DESCRIPTION}
\item{\code{isBioEroder}}{integer COLUMN_DESCRIPTION}
\item{\code{isExternal}}{integer COLUMN_DESCRIPTION}
\item{\code{isVertebrate}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
functionalgroups
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
