% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_salmonpredators.R
\name{get_salmonpredators}
\alias{get_salmonpredators}
\title{Get salmon predators}
\usage{
get_salmonpredators(ppreymatrix, functionalgroups)
}
\arguments{
\item{functionalgroups}{}
}
\value{
predgroups
}
\description{
Get salmon predators
}
