% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shapefile.R
\name{get_shapefile}
\alias{get_shapefile}
\title{download shapefiles}
\usage{
get_shapefile(eachfilename, shapefile.name)
}
\arguments{
\item{eachfilename, }{a string}

\item{shapefile.name, }{a string}
}
\description{
Code to plot survival of multiple AMPS versions
}
\author{
Hem Nalini Morzaria-Luna, hmorzarialuna_gmail.com February 2002
}
