% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-groupguilds.R
\docType{data}
\name{groupguilds}
\alias{groupguilds}
\title{DATASET_TITLE}
\format{
A data frame with 73 rows and 6 variables:
\describe{
\item{\code{code}}{character Functional group code}
\item{\code{index}}{integer Index}
\item{\code{name}}{character Functional group name}
\item{\code{long_name}}{character Functional group long name}
\item{\code{group_type}}{character Functional group name Atlantis group type}
\item{\code{guild}}{character Functional group guild}
}
}
\usage{
groupguilds
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
