% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trajectories.R
\name{make_trajectories}
\alias{make_trajectories}
\title{analyze trajectories}
\usage{
make_trajectories(ensemblenumbersagescenarios, salmongroups)
}
\arguments{
\item{ensemblenumbersagescenarios}{}
}
\value{

}
\description{
analyze trajectories
}
