% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dietcheck.R
\name{plot_dietcheck}
\alias{plot_dietcheck}
\title{Plot Dietcheck time series}
\usage{
plot_dietcheck(dietcheck, indsalmoneffect, groupguilds)
}
\arguments{
\item{dietcheck}{total predation by predators on salmon groups for individual scenarios}

\item{groupguilds}{functional groups by guild}

\item{salmoneffect}{predicted scenario effects on salmon}
}
\value{
pred.boxplot
}
\description{
Plot Dietcheck time series
}
\author{
Hem Nalini Morzaria-Luna, hmorzarialuna_gmail.com May 2023
}
