% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dietcheckcum.R
\name{plot_dietcheckcum}
\alias{plot_dietcheckcum}
\title{Plot Dietcheckcum time series}
\usage{
plot_dietcheckcum(dietcheckcum, salmoneffect, predgroups, thiscutoff)
}
\arguments{
\item{dietcheckcum}{total predation by predators on salmon groups for cumulative scenarios}

\item{salmoneffect}{predicted scenario effects on salmon}

\item{groupguilds}{functional groups by guild}
}
\value{
pred.boxplot
}
\description{
Plot Dietcheckcum time series
}
\author{
Hem Nalini Morzaria-Luna, hmorzarialuna_gmail.com May 2023
}
