% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ensemble_survival_scenarios.R
\name{plot_ensemble_survival_scenarios}
\alias{plot_ensemble_survival_scenarios}
\title{plot salmon survival of ensemble models}
\usage{
plot_ensemble_survival_scenarios(
  ensemblenumbersagescenarios,
  salmongroups,
  plotmodels,
  base.survival,
  salmonbybasin,
  scenario.effect
)
}
\arguments{
\item{ensemblenumbersage, }{a data frame}

\item{func.groups, }{a data frame of salmon groups in the model}
}
\value{
salmon.return.nums, survival over time
}
\description{
Code to plot survival of multiple AMPS versions
}
\author{
Hem Nalini Morzaria-Luna, hmorzarialuna_gmail.com February 2022
}
