% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ensemble_survival_scenarios_timeseries.R
\name{plot_ensemble_survival_scenarios_timeseries}
\alias{plot_ensemble_survival_scenarios_timeseries}
\title{plot time series of salmon survival from ensemble models}
\usage{
plot_ensemble_survival_scenarios_timeseries(
  ensemblenumbersagescenarios,
  salmongroups,
  scenario.effect
)
}
\arguments{
\item{ensemblenumbersage, }{a data frame}

\item{func.groups, }{a data frame of salmon groups in the model}
}
\value{

}
\description{
Code to plot survival of multiple models over time
}
\author{
Hem Nalini Morzaria-Luna, hmorzarialuna_gmail.com June 2002
}
