% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_foodweb.R
\name{plot_foodweb}
\alias{plot_foodweb}
\title{plot Atlantis food web}
\usage{
plot_foodweb(ppreymatrix, plot.name)
}
\arguments{
\item{ensemblebiomass, }{a data frame}

\item{plotmodels, }{a list of models to plot}
}
\value{
models.used, data frame of final models plotted
}
\description{
Code to create Atlantis food web from prey matrix that show linkages with 0 & 1

reads square matrix of predators (rows) consuming prey (columns)
}
\details{
Hem Nalini Morzaria-Luna
hmorzarialuna@gmail.com
Code to create Atlantis food web from prey matrix that show linkages with 0 & 1
uses package ggnet2 https://briatte.github.io/ggnet/
}
\author{
Hem Nalini Morzaria-Luna, hmorzarialuna_gmail.com, February 2022
}
