% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interaction_effect.R
\name{plot_interaction_effect}
\alias{plot_interaction_effect}
\title{Plot interaction effect}
\usage{
plot_interaction_effect(
  ensemble.survival,
  ensemble.cum.survival,
  scenariocategories,
  salmnonbybasin
)
}
\arguments{
\item{salmon.colors}{}
}
\value{

}
\description{
Plot interaction effect
}
\examples{

}
