% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predation.R
\name{plot_predation}
\alias{plot_predation}
\title{plot predation}
\usage{
plot_predation(ensemblepredationcum, salmoneffect, salmonbybasin)
}
\arguments{
\item{ensemblenumbersage, }{a data frame}

\item{func.groups, }{a data frame of salmon groups in the model}
}
\value{
salmon.return.nums, survival over time
}
\description{
Code to plot predation changes between scenarios
}
\author{
Hem Nalini Morzaria-Luna, hmorzarialuna_gmail.com November 2022
}
