% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predators.R
\name{plot_predators}
\alias{plot_predators}
\title{plot biomass salmon predators}
\usage{
plot_predators(ensemblebiomass, predgroups, thiscutoff)
}
\arguments{
\item{ensemblebiomass}{biomass individual scenarios}

\item{predgroups}{Salmon predator groups in guilds}
}
\value{
biomass plots of salmon predators
}
\description{
Code to plot biomass of salmon predators in individual scenarios
}
\author{
Hem Nalini Morzaria-Luna, hmorzarialuna_gmail.com June 2023
}
