% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predatorscum.R
\name{plot_predatorscum}
\alias{plot_predatorscum}
\title{plot biomass salmon predators in cumulative scenarios}
\usage{
plot_predatorscum(ensemblebiomasscum, predgroups, thiscutoff)
}
\arguments{
\item{ensemblebiomasscum}{biomass cumulative scenarios}

\item{predgroups}{Salmon predator groups in guilds}
}
\value{
biomass plots of salmon predators
}
\description{
Code to plot biomass of salmon predators in cumulative scenarios
}
\author{
Hem Nalini Morzaria-Luna, hmorzarialuna_gmail.com February 2023
}
