% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_salmon_abundance.R
\name{plot_salmon_abundance}
\alias{plot_salmon_abundance}
\title{plot salmon abundance from Lossee et al}
\usage{
plot_salmon_abundance(salmon.abundance)
}
\arguments{
\item{salmon.abundance, }{a data frame}
}
\value{
salmon.plot
}
\description{
Code to plot survival of multiple AMPS versions
}
\author{
Hem Nalini Morzaria-Luna, hmorzarialuna_gmail.com February 2002
}
