% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_watage.R
\name{plot_watage}
\alias{plot_watage}
\title{Plot weight-at-age for individual scenarios}
\usage{
plot_watage(ensemblewageind, salmonbybasin, salmongroups, thiscutoff)
}
\arguments{
\item{salmongroups}{}
}
\value{
wage.boxplot
}
\description{
Plot weight-at-age for individual scenarios
}
\examples{
plot_predation(ensemblewageind, salmonbybasin, salmongroups, thiscutoff=100)
}
