% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_watage_cum.R
\name{plot_watage_cum}
\alias{plot_watage_cum}
\title{Plot weight-at-age for cumulative scenarios}
\usage{
plot_watage_cum(
  ensemblewagecum,
  salmonbybasin,
  salmongroups,
  salmoneffect,
  thiscutoff
)
}
\arguments{
\item{salmongroups}{}
}
\value{
wage.boxplot
}
\description{
Plot weight-at-age for cumulative scenarios
}
\examples{
plot_predation(ensemblewagecum, salmonbybasin, salmongroups, salmoneffect,thiscutoff=20)
}
