% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-predgroups.R
\docType{data}
\name{predgroups}
\alias{predgroups}
\title{DATASET_TITLE}
\format{
A data frame with 34 rows and 6 variables:
\describe{
\item{\code{Predator}}{character COLUMN_DESCRIPTION}
\item{\code{Code}}{character COLUMN_DESCRIPTION}
\item{\code{name}}{character COLUMN_DESCRIPTION}
\item{\code{GroupType}}{character COLUMN_DESCRIPTION}
\item{\code{isExternal}}{integer COLUMN_DESCRIPTION}
\item{\code{guild}}{character COLUMN_DESCRIPTION}
}
}
\usage{
predgroups
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
