% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmon_colors.R
\docType{data}
\name{salmon.colors}
\alias{salmon.colors}
\title{Vector of custom colors}
\format{
An object of class \code{character} of length 21.
}
\usage{
salmon.colors
}
\description{
Vector of custom colors
}
\examples{
salmon.colors[1]

}
\keyword{datasets}
