% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmon_colors.R
\name{salmon_colors}
\alias{salmon_colors}
\title{Vector of custom salmon colors}
\usage{
salmon_colors(...)
}
\arguments{
\item{...}{}
}
\value{
salmon.colors
}
\description{
Vector of custom salmon colors
}
