% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-salmonbybasin.R
\docType{data}
\name{salmonbybasin}
\alias{salmonbybasin}
\title{DATASET_TITLE}
\format{
A data frame with 21 rows and 6 variables:
\describe{
\item{\code{Code}}{character salmon group code}
\item{\code{name}}{character salmon group name}
\item{\code{longname}}{character salmon group long name}
\item{\code{basin}}{character Puget Sound basin}
\item{\code{geo_order}}{integer geographic order}
\item{\code{salmon_genus}}{character salmon genus}
}
}
\usage{
salmonbybasin
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
