% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-salmoneffect.R
\docType{data}
\name{salmoneffect}
\alias{salmoneffect}
\title{Expected salmon effect}
\format{
A data frame with 6 rows and 3 variables:
\describe{
\item{\code{scenario_name}}{character Scenario name}
\item{\code{scenario_var}}{character Salmon effect scenario}
\item{\code{salmon_effect}}{character Expected salmon effect}
}
}
\usage{
salmoneffect
}
\description{
Expected salmon effects from scenarios
}
\details{
DETAILS
}
\keyword{datasets}
