% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-scenariocategories.R
\docType{data}
\name{scenariocategories}
\alias{scenariocategories}
\title{scenariocategories}
\format{
A data frame with 9 rows and 2 variables:
\describe{
\item{\code{scenario_name}}{character Name of scenario}
\item{\code{scenario_category}}{character Category as top-down or bottom-up}
}
}
\usage{
scenariocategories
}
\description{
Categories of scenarios as top-down or bottom-up
}
\details{
categories to summarize survival values and compare with cumulative scenarios
}
\keyword{datasets}
