% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dietcheckcum.R
\docType{data}
\name{dietcheckcum}
\alias{dietcheckcum}
\title{Salmon consumption}
\format{
A data frame with 691200 rows and 6 variables:
\describe{
\item{\code{Time}}{double timestep}
\item{\code{Predator}}{character predator}
\item{\code{model_ver}}{integer model version}
\item{\code{scenario_name}}{character scenario name}
\item{\code{scenario_var}}{character salmon effect scenario}
\item{\code{tot_mort}}{double total consumption}
}
}
\usage{
dietcheckcum
}
\description{
Consumption on salmon based on dietcheck for cumulative scenarios
}
\details{
DETAILS
}
\keyword{datasets}
