% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ensemblebiomass.R
\docType{data}
\name{ensemblebiomass}
\alias{ensemblebiomass}
\title{Biomass in individual scenarios}
\format{
A data frame with 398268 rows and 8 variables:
\describe{
\item{\code{Code}}{character species code}
\item{\code{biomass}}{double biomass in mt}
\item{\code{name}}{character species name}
\item{\code{longname}}{character species long name}
\item{\code{Year}}{double year}
\item{\code{model_ver}}{integer model ensemble version}
\item{\code{scenario_name}}{character scenario name}
\item{\code{scenario_var}}{character senario variation}
}
}
\usage{
ensemblebiomass
}
\description{
Biomass in individual salmon survival scenarios
}
\details{
DETAILS
}
\keyword{datasets}
