% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ensemblebiomasscum.R
\docType{data}
\name{ensemblebiomasscum}
\alias{ensemblebiomasscum}
\title{Biomass cumulative scenarios}
\format{
A data frame with 138528 rows and 8 variables:
\describe{
\item{\code{Code}}{character Functional group code}
\item{\code{biomass}}{double Biomass metric tons}
\item{\code{name}}{character Functional group name}
\item{\code{longname}}{character Functional group long name}
\item{\code{Year}}{double Year}
\item{\code{model_ver}}{integer Model ensemble member}
\item{\code{scenario_name}}{character Scenario name}
\item{\code{scenario_var}}{character Scenario expected effect on salmon}
}
}
\usage{
ensemblebiomasscum
}
\description{
Biomass for all groups in cumulative scenarios
}
\details{
DETAILS
}
\keyword{datasets}
