% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ensemblebiomasscum.R
\docType{data}
\name{ensemblebiomasscum}
\alias{ensemblebiomasscum}
\title{DATASET_TITLE}
\format{
A data frame with 536352 rows and 8 variables:
\describe{
\item{\code{Code}}{character functional group code}
\item{\code{biomass}}{double biomass}
\item{\code{name}}{character functional group name}
\item{\code{longname}}{character functional group long name}
\item{\code{Year}}{double year}
\item{\code{model_ver}}{integer model ensemble version}
\item{\code{scenario_name}}{character scenario name}
\item{\code{scenario_var}}{character scenario ver}
}
}
\usage{
ensemblebiomasscum
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
