% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ensemblenumberscum.R
\docType{data}
\name{ensemblenumberscum}
\alias{ensemblenumberscum}
\title{DATASET_TITLE}
\format{
A data frame with 705600 rows and 14 variables:
\describe{
\item{\code{nums}}{double numbers}
\item{\code{agename}}{character age class}
\item{\code{group}}{character functional group}
\item{\code{age}}{integer age}
\item{\code{Code}}{character functional group code}
\item{\code{years_away}}{integer years away}
\item{\code{longname}}{character functional group long name}
\item{\code{NumCohorts}}{integer number of cohorts}
\item{\code{timestep}}{integer timestep}
\item{\code{year}}{double year}
\item{\code{year_no}}{integer year number}
\item{\code{model_ver}}{integer model ensemble version}
\item{\code{scenario_name}}{character scenario name}
\item{\code{scenario_var}}{integer scenario variation}
}
}
\usage{
ensemblenumberscum
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
