% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-indsalmoneffect.R
\docType{data}
\name{indsalmoneffect}
\alias{indsalmoneffect}
\title{DATASET_TITLE}
\format{
A data frame with 18 rows and 3 variables:
\describe{
\item{\code{scenario_name}}{character COLUMN_DESCRIPTION}
\item{\code{scenario_var}}{character COLUMN_DESCRIPTION}
\item{\code{salmon_effect}}{character COLUMN_DESCRIPTION}
}
}
\usage{
indsalmoneffect
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
