% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_forcing.R
\name{make_forcing}
\alias{make_forcing}
\title{make forcing files for salmon survival}
\usage{
make_forcing(
  thislistitem,
  ncfile,
  func.groups,
  num.forcing.years,
  rate.multiplier,
  bash.file,
  force.prm,
  salmongroups
)
}
\arguments{
\item{ncfile, }{AMPS}

\item{func.groups, }{a data frame of salmon groups in the model}
}
\value{
salmon.return.nums, survival over time
}
\description{
Code to modify AMPS nc file to create forcing files for sensitivity simulations

function to create forcing files, force prm and bash file
}
\author{
Hem Nalini Morzaria-Luna, hmorzarialuna_gmail.com February 2002
}
