% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map.R
\name{make_map}
\alias{make_map}
\title{Mapping Atlantis polygons}
\usage{
make_map(file.name, get_shapefile)
}
\arguments{
\item{shape.file}{which shape file to map}
}
\value{
model.map
}
\description{
Functions to plot shapefile
}
\details{
INPUT: 1) An Atlantis model polygon shape file

OUTPUT: 1) Map of Atlantis polygons

Used code from https://www.r-spatial.org/r/2018/10/25/ggplot2-sf.html
}
\author{
Hem Nalini Morzaria-Luna, hmorzarialuna@gmail.com
}
