% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ensemble_cum_survival_scenarios.R
\name{plot_ensemble_cum_survival_scenarios}
\alias{plot_ensemble_cum_survival_scenarios}
\title{plot salmon survival of cumulative impact scenarios}
\usage{
plot_ensemble_cum_survival_scenarios(
  ensemblenumberscum,
  salmongroups,
  base.cum.survival,
  salmonbybasin,
  salmoneffect,
  salmon.colors
)
}
\arguments{
\item{ensemblenumbersage, }{a data frame}

\item{func.groups, }{a data frame of salmon groups in the model}
}
\value{
salmon_cum_return_nums.csv, survival over time
}
\description{
Code to plot survival of multiple AMPS versions
}
\author{
Hem Nalini Morzaria-Luna, hmorzarialuna_gmail.com February 2022
}
