% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ensemblebiomass.R
\name{plot_ensemblebiomass}
\alias{plot_ensemblebiomass}
\title{plot biomass of ensemble models}
\usage{
plot_ensemblebiomass(ensemblebiomass)
}
\arguments{
\item{ensemblebiomass, }{a data frame}

\item{plotmodels, }{a list of models to plot}
}
\value{
models.used, data frame of final models plotted
}
\description{
Code to plot biomass of multiple AMPS versions
}
\author{
Hem Nalini Morzaria-Luna, hmorzarialuna_gmail.com, February 2022
}
