% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predators.R
\name{plot_predators}
\alias{plot_predators}
\title{plot biomass salmon predators}
\usage{
plot_predators(ensemblebiomasscum, predgroups)
}
\arguments{
\item{ensemblebiomasscum}{biomass cumulative scenarios}

\item{predgroups}{Salmon predator groups in guilds}
}
\value{
biomass plots of salmon predators
}
\description{
Code to plot biomass of salmon predators in cumulative scenarios
}
\author{
Hem Nalini Morzaria-Luna, hmorzarialuna_gmail.com February 2023
}
