% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-predgroups.R
\docType{data}
\name{predgroups}
\alias{predgroups}
\title{Consumer groups}
\format{
A data frame with 34 rows and 6 variables:
\describe{
\item{\code{Predator}}{character Consumer name}
\item{\code{Code}}{character Consumer species code}
\item{\code{name}}{character Consumer name}
\item{\code{GroupType}}{characterGroup type}
\item{\code{isExternal}}{integer Does group have external reproduction}
\item{\code{guild}}{character Species guild}
}
}
\usage{
predgroups
}
\description{
Consumer groups in the AMPS
}
\details{
DETAILS
}
\keyword{datasets}
