% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-predgroups.R
\docType{data}
\name{predgroups}
\alias{predgroups}
\title{DATASET_TITLE}
\format{
A data frame with 23 rows and 6 variables:
\describe{
\item{\code{Code}}{character species code}
\item{\code{group_no}}{integer group number}
\item{\code{guild}}{character species guild}
\item{\code{atlantis_name}}{character species model name}
\item{\code{long_name}}{character species long name}
\item{\code{NumCohorts}}{integer number of cohorts}
}
}
\usage{
predgroups
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
