% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-salmonbybasin.R
\docType{data}
\name{salmonbybasin}
\alias{salmonbybasin}
\title{DATASET_TITLE}
\format{
A data frame with 21 rows and 6 variables:
\describe{
\item{\code{Code}}{character COLUMN_DESCRIPTION}
\item{\code{name}}{character COLUMN_DESCRIPTION}
\item{\code{longname}}{character COLUMN_DESCRIPTION}
\item{\code{basin}}{character COLUMN_DESCRIPTION}
\item{\code{geo_order}}{integer COLUMN_DESCRIPTION}
\item{\code{salmon_genus}}{character COLUMN_DESCRIPTION}
}
}
\usage{
salmonbybasin
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
