% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-salmonbybasin.R
\docType{data}
\name{salmonbybasin}
\alias{salmonbybasin}
\title{Salmon by basin}
\format{
A data frame with 21 rows and 6 variables:
\describe{
\item{\code{Code}}{character species code}
\item{\code{name}}{character species name}
\item{\code{longname}}{character species long name}
\item{\code{basin}}{character Puget Sound basin}
\item{\code{geo_order}}{integer Geographical order}
\item{\code{salmon_genus}}{character Salmon genus}
}
}
\usage{
salmonbybasin
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
