% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-salmoneffect.R
\docType{data}
\name{salmoneffect}
\alias{salmoneffect}
\title{DATASET_TITLE}
\format{
A data frame with 6 rows and 3 variables:
\describe{
\item{\code{scenario_name}}{character scenario name}
\item{\code{scenario_var}}{character scenario variation}
\item{\code{salmon_effect}}{character salmon effect}
}
}
\usage{
salmoneffect
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
